move shared API contracts out of pages/api so front-api no longer imports Next handlers
