+++ front/lib/api/assistant/user_relation.ts
@@ -3,6 +3,21 @@ import type { Authenticator } from "@app/lib/auth";
 import { AgentUserRelationModel } from "@app/lib/models/agent/agent";
 import type { Result } from "@app/types/shared/result";
 import { Err, Ok } from "@app/types/shared/result";
+import { z } from "zod";
+
+export type PostAgentUserFavoriteResponseBody = {
+  agentId: string;
+  userFavorite: boolean;
+};
+
+export const PostAgentUserFavoriteRequestBodySchema = z.object({
+  agentId: z.string(),
+  userFavorite: z.boolean(),
+});
+
+export type PostAgentUserFavoriteRequestBody = z.infer<
+  typeof PostAgentUserFavoriteRequestBodySchema
+>;
 
 export async function setAgentUserFavorite({
   auth,
