+++ front/lib/api/invitation.ts
@@ -31,14 +31,46 @@ import type {
   UserType,
   WorkspaceType,
 } from "@app/types/user";
+import { ActiveRoleSchema } from "@app/types/user";
 import sgMail from "@sendgrid/mail";
 import { escape } from "html-escaper";
 import type { Transaction } from "sequelize";
 import { Op } from "sequelize";
+import { z } from "zod";
 
 import { MembershipInvitationResource } from "../resources/membership_invitation_resource";
 
 const EMAIL_CONCURRENCY = 8;
+
+export type GetWorkspaceInvitationsResponseBody = {
+  invitations: MembershipInvitationType[];
+};
+
+export const PostInvitationRequestBodySchema = z.array(
+  z.object({
+    email: z.string(),
+    role: ActiveRoleSchema,
+  })
+);
+
+export type PostInvitationRequestBody = z.infer<
+  typeof PostInvitationRequestBodySchema
+>;
+
+export type PostInvitationResponseBody = {
+  success: boolean;
+  email: string;
+  error_message?: string;
+}[];
+
+export type PostMemberInvitationsResponseBody = {
+  invitation: MembershipInvitationType;
+};
+
+export const PostMemberInvitationBodySchema = z.object({
+  status: z.enum(["revoked", "pending"]),
+  initialRole: ActiveRoleSchema,
+});
 
 export async function getInvitation(
   auth: Authenticator,
