+++ front/lib/api/skills/schemas.ts
@@ -1,0 +1,9 @@
+import { z } from "zod";
+
+// Schema for knowledge attached to a skill.
+export const AttachedKnowledgeSchema = z.object({
+  dataSourceViewId: z.string(),
+  nodeId: z.string(),
+  spaceId: z.string(),
+  title: z.string(),
+});
