+++ front/pages/api/w/[wId]/invitations/[iId]/index.ts
@@ -6,24 +6,14 @@ import {
   getAuditLogContext,
 } from "@app/lib/api/audit/workos_audit";
 import { withSessionAuthenticationForWorkspace } from "@app/lib/api/auth_wrappers";
+import type { PostMemberInvitationsResponseBody } from "@app/lib/api/invitation";
+import { PostMemberInvitationBodySchema } from "@app/lib/api/invitation";
 import type { Authenticator } from "@app/lib/auth";
 import { MembershipInvitationResource } from "@app/lib/resources/membership_invitation_resource";
 import { apiError } from "@app/logger/withlogging";
 import type { WithAPIErrorResponse } from "@app/types/error";
-import type { MembershipInvitationType } from "@app/types/membership_invitation";
-import { ActiveRoleSchema } from "@app/types/user";
 import type { NextApiRequest, NextApiResponse } from "next";
-import { z } from "zod";
 import { fromError } from "zod-validation-error";
-
-export type PostMemberInvitationsResponseBody = {
-  invitation: MembershipInvitationType;
-};
-
-export const PostMemberInvitationBodySchema = z.object({
-  status: z.enum(["revoked", "pending"]),
-  initialRole: ActiveRoleSchema,
-});
 
 async function handler(
   req: NextApiRequest,
